Rewrite partition handling, broken by fdisk
The expected behavior for (s)fdisk when the disk has no partitions is
to return a non-zero code. However, in some cases, fdisk things it sees
a dos type partition table and when it is looking at a filesystem. This
is obviously an issue, so we remove it and use partx's output instead.